Package artifacts defined in template Metadata
This generalizes the packaging functionality used by 'aws cloudformation
package' so that s3 objects can be used wherever needed in a template.
Makes it possible to retrieve:

* an s3 url
* an https url
* the key for the resulting object
* the bucket for all uploaded objects

All four of may be necessary at different times, for instance a Lambda
function needs the key/bucket, while the 'source' for a
'AWS::CloudFormation::Init' file wants an https url. Currently passes
packaged values to the template by setting them as 'Default'.